requirements.txt
anthropic
httpx
supabase
cohere
python-dotenv
pydantic
eth-account
cryptography
pynacl